[Breaking Change] Fix some patches overriding the wrong patch method.
Also make the bouncer one final so that this doesn't happen again. This
will break anyone with custom patches that override it, but I don't
think this is likely to happen.

The override change shouldn't affect anything outside of the find
operation, which ironically cannot be tested (as move and copy cannot
traverse an element's parent from inside a find patch.)