Fix zero-prefixed numbers in paths being treated as numbers.
I'm not exactly sure what past me was going on about with the comments
about json parsing, since this test doesn't actually have anything to do
with that? If anything, it points out a bug in the path parsing code
(which has nothing to do with json itself). Regardless, this has now
been fixed.